removed entrypoint. just use commands in the compose file
When running from the compose file the container did not print any
logs. Switching to python unbuffered mode (-u) and removing the
exec & eval helped. (Although I'm not sure about if I should use
unbuffered mode or the tty:true option in the compose file. Other
contianers never had the problem that stdout was not printed)

There was one bug with the SALT parameter (no quotes) and I don't
understand why the exec $(eval ...) construction was used in the
first place. exec would make sense if you don't want to spawn another
process but no idea about the eval. Maybe it was in order to pass the
--args to the pyhton process but that worked even without eval.

So I just use commands instead, which get passed to the entrypoint.
No fiddling around with environment variables necessary.